# api/src/main/resources/ApiContract.yaml
swagger: '2.0'
info:
  description: This is api rest contract
  version: 1.0.0
  title: Swagger FlashPet
basePath: /v1
tags:
  - name: company
    description: Everything about your Pets
  - name: auth
    description: Auth service
  - name: owner
    description: Owner service
schemes:
  - https
  - http
paths:
  /companies:
    post:
      tags:
        - company
      summary: Add a new company
      description: ''
      operationId: addCompany
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Company'
      responses:
        '201':
          description: Company saved
        '405':
          description: Invalid input
          schema:
            $ref: '#/definitions/ModelApiResponse'
    put:
      tags:
        - company
      summary: Update an existing Company
      description: ''
      operationId: updateCompany
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Company'
      responses:
        '200':
          description: Company updated
        '400':
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/ModelApiResponse'
        '404':
          description: Company not found
          schema:
            $ref: '#/definitions/ModelApiResponse'
        '405':
          description: Validation exception
          schema:
            $ref: '#/definitions/ModelApiResponse'
    get:
      tags:
        - company
      parameters:
        - in: query
          name: service_description
          type: string
        - in: query
          name: price_category
          type: array
          items:
            enum:
              - low_000_300
              - midlle_300_600
              - high_600_up
        - in: query
          name: service_type
          type: array
          items:
            type: string
        - in: query
          name: latitude
          type: number
          format: double
        - in: query
          name: longitude
          type: number
          format: double
        - in: query
          name: day
          type: array
          items:
            type: string
        - in: query
          name: rateInit
          type: integer
          format: int32
        - in: query
          name: rateFinal
          type: integer
          format: int32
      summary: Finds all companies
      operationId: findCompanies
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              scrollId:
                type: string
              resultMatchs:
                type: string
              companies:
                type: array
                items:
                  $ref: '#/definitions/Get_Company'
        '400':
          description: Invalid status value
          schema:
            $ref: '#/definitions/ModelApiResponse'
  /companies/{scroll}:
    get:
      tags:
        - company
      parameters:
        - in: path
          name: scroll
          type: string
          required: true
      summary: Finds all companies
      operationId: findCompaniesByScroll
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              scrollId:
                type: string
              resultMatchs:
                type: string
              companies:
                type: array
                items:
                  $ref: '#/definitions/Get_Company'
        '400':
          description: Invalid status value
          schema:
            $ref: '#/definitions/ModelApiResponse'
  /companies/{companyId}/rates:
    post:
      tags:
        - company
      parameters:
        - in: path
          name: companyId
          type: string
          required: true
        - in: query
          name: cache_code
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/IndividualRate'
      summary: Add a rate for especific company
      operationId: addRate
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
          schema:
            $ref: '#/definitions/ModelApiResponse'
  /companies/{companyId}/rates/generate:
    get:
      tags:
        - company
      parameters:
        - in: path
          name: companyId
          type: string
          required: true
      summary: get cache code of rate
      operationId: getRate
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/CacheRate'
        '400':
          description: Invalid status value
          schema:
            $ref: '#/definitions/ModelApiResponse'
  /authenticate:
    post:
      tags:
        - auth
      summary: Endpoint to get jwt token
      description: ''
      operationId: auth
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Jwt_Request'
      responses:
        '200':
          description: Jwt generated
          schema:
            $ref: '#/definitions/Jwt_Response'
        '405':
          description: Invalid input
          schema:
            $ref: '#/definitions/ModelApiResponse'
  /owners:
    get:
      tags:
        - owner
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
      summary: Finds all companies
      operationId: findCompaniesByOwner
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              scrollId:
                type: string
              resultMatchs:
                type: string
              companies:
                type: array
                items:
                  $ref: '#/definitions/Company'
        '400':
          description: Invalid status value
          schema:
            $ref: '#/definitions/ModelApiResponse'
definitions:
  Company:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      address:
        type: string
        format: string
      latitude:
        type: number
        format: double
      longitude:
        type: number
        format: double
      services:
        type: array
        items:
          $ref: '#/definitions/Service'
      owner:
        $ref: '#/definitions/Owner'
      days:
        type: object
        additionalProperties:
          type: object
          properties:
            time:
              $ref: '#/definitions/Time'
      contact:
        type: array
        items:
          $ref: '#/definitions/Contact'
      rate:
        $ref: '#/definitions/Rate'
      logo:
        type: string
        format: string
      carrosel:
        type: array
        items:
          type: string
          format: string
  Get_Company:
    type: object
    properties:
      name:
        type: string
        format: string
      address:
        type: string
        format: string
      latitude:
        type: number
        format: double
      longitude:
        type: number
        format: double
      services:
        type: array
        items:
          $ref: '#/definitions/Get_Service'
      days:
        type: object
        additionalProperties:
          type: object
          properties:
            time:
              $ref: '#/definitions/Time'
      contact:
        type: array
        items:
          $ref: '#/definitions/Contact'
      rate:
        $ref: '#/definitions/Rate'
      logo:
        type: string
        format: string
      carrosel:
        type: array
        items:
          type: string
          format: string
  Get_Service:
    type: object
    properties:
      description:
        type: string
        example: Service to shower your pet
      price:
        type: number
        format: double
      service_type:
        type: string
      image:
        type: string
        format: string
  Service:
    type: object
    properties:
      id:
        type: string
        format: string
      description:
        type: string
        example: Service to shower your pet
      price:
        type: number
        format: double
      service_type:
        type: string
      image:
        type: string
        format: string
  Owner:
    type: object
    properties:
      id:
        type: string
        format: string
      username:
        type: string
        format: string
      password:
        type: string
        format: string
      email:
        type: string
        format: string
      cnpj:
        type: string
        format: string
  Time:
    type: object
    properties:
      start:
        type: string
        format: string
      end:
        type: string
        format: string
  Contact:
    type: object
    properties:
      ContactType:
        type: string
      number:
        type: string
        format: string
  Rate:
    type: object
    properties:
      avg:
        type: number
        format: double
      individual_rate:
        type: array
        items:
          $ref: '#/definitions/IndividualRate'
  CacheRate:
    type: object
    properties:
      code:
        type: string
        format: string
      expiration:
        type: string
        format: string
  IndividualRate:
    type: object
    properties:
      date:
        type: string
        format: string
      description:
        type: string
        format: string
      value:
        type: integer
        format: int32
  Jwt_Request:
    type: object
    properties:
      username:
        type: string
        format: string
      password:
        type: string
        format: string
  Jwt_Response:
    type: object
    properties:
      token:
        type: string
        format: string
  ModelApiResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string