# domain/src/main/java/com/senac/domain/entity/Contact.java
package com.senac.domain.entity;

import lombok.*;
import org.springframework.data.annotation.Id;
import org.springframework.data.elasticsearch.annotations.Document;
import org.springframework.data.elasticsearch.annotations.Field;
import org.springframework.data.elasticsearch.annotations.FieldType;

@Builder
@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
@Document(indexName = "contact_index")
public class Contact {
    @Id
    private String id;

    @Field(type = FieldType.Text)
    private String contactType;

    @Field(type = FieldType.Text)
    private String number;
}